# Estimate length using logarithms