I'm not sure if it's faster (we really need a benchmark suite) but it allows Clang to vectorize the loop.

Check the assembly diff:
https://www.diffchecker.com/Ga73O0D9
The original is your version. The diff is this PR.