@thedrow 2016-08-31
These are the results of my Python benchmark suite:
# Original
```
------------------------------------------------------------------------- benchmark 'encode': 7 tests --------------------------------------------------------------------------
Name (time in us)            Min                     Max               Mean                StdDev            Median               IQR            Outliers(*)  Rounds  Iterations
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_encode_cffi[0]       2.5034 (1.0)       44,866.5857 (932.99)    4.4521 (1.11)       226.1240 (122.67)   2.7895 (1.0)      0.2146 (1.12)          6;4762   39946          10
test_encode_cffi[4]       2.5988 (1.04)      65,355.5155 (>1000.0)   5.1584 (1.29)       335.7273 (182.13)   2.8133 (1.01)     0.1907 (1.0)           7;5241   38480          10
test_encode_cffi[12]      3.3855 (1.35)          83.2081 (1.73)      4.0417 (1.01)         2.1331 (1.16)     3.6001 (1.29)     0.2861 (1.50)       1059;3825   29538          10
test_encode_cffi[8]       3.3855 (1.35)          90.0984 (1.87)      4.0407 (1.01)         2.3391 (1.27)     3.6001 (1.29)     0.2861 (1.50)       1121;3426   29538          10
test_encode_cffi[24]      3.4809 (1.39)          48.0890 (1.0)       3.9948 (1.0)          1.8433 (1.0)      3.6001 (1.29)     0.1907 (1.00)        934;4389   29538          10
test_encode_cffi[48]      3.5763 (1.43)          67.9016 (1.41)      4.1825 (1.05)         1.8725 (1.02)     3.7193 (1.33)     0.3099 (1.63)       1012;3494   27963          10
test_encode_cffi[120]     5.4836 (2.19)     574,469.5663 (>1000.0)  12.8808 (3.22)     1,902.5083 (>1000.0)  5.9605 (2.14)     0.5960 (3.13)         3;10466   91181           2
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```
# This PR
```
------------------------------------------------------------------------- benchmark 'encode': 7 tests --------------------------------------------------------------------------
Name (time in us)            Min                     Max               Mean                StdDev            Median               IQR            Outliers(*)  Rounds  Iterations
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_encode_cffi[0]       2.5988 (1.0)       40,944.9100 (>1000.0)   4.6181 (1.16)       214.1275 (129.17)   2.9087 (1.0)      0.2146 (1.12)          5;5257   37118          10
test_encode_cffi[4]       2.6941 (1.04)      60,787.4155 (>1000.0)   5.2234 (1.32)       318.1371 (191.92)   2.9087 (1.0)      0.2861 (1.50)          7;3488   37118          10
test_encode_cffi[8]       3.3855 (1.30)       1,041.8892 (28.23)     4.0480 (1.02)         6.3347 (3.82)     3.6001 (1.24)     0.2146 (1.12)        415;2717   29538          10
test_encode_cffi[12]      3.4809 (1.34)         178.2179 (4.83)      4.0240 (1.01)         2.0864 (1.26)     3.6001 (1.24)     0.2146 (1.12)        850;3869   29538          10
test_encode_cffi[24]      3.4809 (1.34)          36.9072 (1.0)       3.9684 (1.0)          1.6577 (1.0)      3.6001 (1.24)     0.1907 (1.0)         881;3756   28729          10
test_encode_cffi[48]      3.5763 (1.38)          85.9976 (2.33)      4.0857 (1.03)         1.8316 (1.10)     3.6955 (1.27)     0.2146 (1.12)        827;3079   27963          10
test_encode_cffi[120]     4.8876 (1.88)     559,006.9294 (>1000.0)  12.4656 (3.14)     1,761.0005 (>1000.0)  5.4836 (1.89)     0.4768 (2.50)         4;11098  102301           2
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

It looks statistically insignificant but the Python benchmark suite involves a lot of things such as GC and converting from Python types to C types and back.