@thedrow 2016-08-31
With the new commit:
```
----------------------------------------------------------------------- benchmark 'encode': 7 tests ------------------------------------------------------------------------
Name (time in us)            Min                    Max              Mean              StdDev            Median               IQR            Outliers(*)  Rounds  Iterations
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_encode_cffi[0]       2.4796 (1.0)      43,265.3904 (>1000.0)  4.3907 (1.12)     218.4513 (133.35)   2.7895 (1.0)      0.2146 (1.0)           6;4074   40330          10
test_encode_cffi[4]       2.5988 (1.05)     61,198.4968 (>1000.0)  5.0504 (1.29)     314.2737 (191.85)   2.8133 (1.01)     0.2146 (1.00)          5;3943   38480          10
test_encode_cffi[12]      3.3855 (1.37)         38.2185 (1.0)      3.9069 (1.0)        1.6382 (1.0)      3.6001 (1.29)     0.2146 (1.00)        832;3781   29538          10
test_encode_cffi[8]       3.3855 (1.37)     74,884.5816 (>1000.0)  7.0460 (1.80)     445.2695 (271.81)   3.6955 (1.32)     0.3815 (1.78)          4;2168   28729          10
test_encode_cffi[24]      3.4809 (1.40)         39.7205 (1.04)     3.9902 (1.02)       1.6718 (1.02)     3.6001 (1.29)     0.2146 (1.00)       1029;3019   28729          10
test_encode_cffi[48]      3.5048 (1.41)         60.9159 (1.59)     4.1110 (1.05)       1.6783 (1.02)     3.6955 (1.32)     0.2146 (1.0)        1095;3219   28533          10
test_encode_cffi[120]     5.4836 (2.21)        766.5157 (20.06)    6.4433 (1.65)       6.6917 (4.08)     5.9605 (2.14)     0.5960 (2.78)       929;14965   91181           2
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```