@tzvetkoff 2016-08-31
Well, according to theory, the correct one (still using a loop for all numbers) should be:

``` c
for (i = 0, j = 1 /* null terminator */ ; i < numbers_count; ++i) {
    j += ceil(log2(numbers[i] + 1) / log2(hashids->alphabet_length)) + 1 /* length + separator and/or the 1 extra byte the algorithm needs */;
}
```

This, however, does not handle 2 edge cases:
 - `0`, since `log(1)` is always `0` -- in this case it returns 1 byte shorter buffer size.
- `18446744073709551615` (`0xFFFFFFFFFFFFFFFF`), since log(0) is `-Infinity`.

A better approach might be:

``` c
for (i = 0, j = 1 /* null terminator */ ; i < numbers_count; ++i) {
    if (numbers[i] == 0) {
        j += 2;
    } else if (numbers[i] == 0xFFFFFFFFFFFFFFFFull) {
        j += ceil(log2(numbers[i]) / log2(hashids->alphabet_length)) + 1;
    } else {
        j += ceil(log2(numbers[i] + 1) / log2(hashids->alphabet_length)) + 1;
    }
}
```

Still, having `if`s should be better than doing the whole loopy loop all over.
I'll write a few tests for this (actually, I'll add more tests, a suite for each function - `make check` made that easy)