@tzvetkoff 2016-09-13
OK, I've implemented my approach so we never allocate more than we actually need.